# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorEnvelope {
  error: string;
}

export interface Product {
  slug: string;
  name: string;
  priceCents: number;
  currency: string;
  paymentNetwork: string;
  tagline: string;
  description: string;
  features: string[];
}

export interface PaymentSubmission {
  network: string;
  amountCents: number;
  proofObjectPath: string;
  submittedAt: string;
}

export interface Order {
  id: number;
  product: Product;
  status: string;
  deliveryStatus: string;
  licenseKey?: string | null;
  paymentSubmission?: PaymentSubmission | null;
  createdAt: string;
  updatedAt: string;
}

export type ReviewOrder = Order & {
  userId: string;
  customerEmail: string;
  proofObjectPath?: string;
};

export interface ReviewVerificationInput {
  note?: string;
}

export interface OrderInput {
  /** @minLength 1 */
  productSlug: string;
}

export interface PaymentProofInput {
  network: string;
  /** @minimum 1 */
  amountCents: number;
  /** @minLength 1 */
  proofObjectPath: string;
  message?: string;
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: UploadUrlRequest;
}