# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: catalog
    description: Forex Gobler EA catalogue and customer orders
  - name: Storage
    description: Object storage upload and serving endpoints.
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /products:
    get:
      operationId: listProducts
      tags: [catalog]
      summary: List available Forex Gobler products
      responses:
        "200":
          description: Product catalogue
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /orders:
    get:
      operationId: listOrders
      tags: [catalog]
      summary: List the signed-in customer's orders
      responses:
        "200":
          description: Customer orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
        "401":
          description: Sign-in required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    post:
      operationId: createOrder
      tags: [catalog]
      summary: Start an order for a Forex Gobler product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Invalid product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Sign-in required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /orders/{id}/payment-proof:
    post:
      operationId: submitPaymentProof
      tags: [catalog]
      summary: Submit screenshot proof for an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentProofInput"
      responses:
        "200":
          description: Payment proof submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Invalid proof
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Sign-in required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /review/orders:
    get:
      operationId: listReviewOrders
      tags: [catalog]
      summary: List orders awaiting payment review
      responses:
        "200":
          description: Orders available to authorized reviewers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReviewOrder"
        "403":
          description: Reviewer access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /review/orders/{id}/verify:
    post:
      operationId: verifyOrderPayment
      tags: [catalog]
      summary: Verify payment and email a unique license
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewVerificationInput"
      responses:
        "200":
          description: Payment verified and license emailed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Order is not ready for verification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Reviewer access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Presigned upload URL generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
        "400":
          description: Missing or invalid required fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Failed to generate upload URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /storage/public-objects/{filePath}:
    get:
      tags: [Storage]
      operationId: getPublicObject
      summary: Serve a public asset
      parameters:
        - name: filePath
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Object content
          content:
            "*/*":
              schema:
                type: string
                format: binary
        "404":
          description: File not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve an uploaded object
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Object content
          content:
            "*/*":
              schema:
                type: string
                format: binary
        "404":
          description: Object not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    Product:
      type: object
      required: [slug, name, priceCents, currency, paymentNetwork, tagline, description, features]
      properties:
        slug:
          type: string
        name:
          type: string
        priceCents:
          type: integer
        currency:
          type: string
        paymentNetwork:
          type: string
        tagline:
          type: string
        description:
          type: string
        features:
          type: array
          items:
            type: string
    PaymentSubmission:
      type: object
      required: [network, amountCents, proofObjectPath, submittedAt]
      properties:
        network:
          type: string
        amountCents:
          type: integer
        proofObjectPath:
          type: string
        submittedAt:
          type: string
          format: date-time
    Order:
      type: object
      required: [id, product, status, deliveryStatus, createdAt, updatedAt]
      properties:
        id:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
        status:
          type: string
        deliveryStatus:
          type: string
        licenseKey:
          oneOf:
            - type: string
            - type: "null"
        paymentSubmission:
          oneOf:
            - $ref: "#/components/schemas/PaymentSubmission"
            - type: "null"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ReviewOrder:
      allOf:
        - $ref: "#/components/schemas/Order"
        - type: object
          required: [userId, customerEmail]
          properties:
            userId:
              type: string
            customerEmail:
              type: string
              format: email
            proofObjectPath:
              type: string
    ReviewVerificationInput:
      type: object
      properties:
        note:
          type: string
    OrderInput:
      type: object
      required: [productSlug]
      properties:
        productSlug:
          type: string
          minLength: 1
    PaymentProofInput:
      type: object
      required: [network, amountCents, proofObjectPath]
      properties:
        network:
          type: string
        amountCents:
          type: integer
          minimum: 1
        proofObjectPath:
          type: string
          minLength: 1
        message:
          type: string
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
        size:
          type: integer
          minimum: 1
        contentType:
          type: string
          minLength: 1
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          format: uri
        objectPath:
          type: string
        metadata:
          $ref: "#/components/schemas/UploadUrlRequest"