# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PaymentSubmission } from './paymentSubmission';
import type { Product } from './product';

export interface Order {
  id: number;
  product: Product;
  status: string;
  deliveryStatus: string;
  licenseKey?: string | null;
  paymentSubmission?: PaymentSubmission | null;
  createdAt: Date;
  updatedAt: Date;
}