# lib/api-zod/src/generated/types/paymentProofInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PaymentProofInput {
  network: string;
  /** @minimum 1 */
  amountCents: number;
  /** @minLength 1 */
  proofObjectPath: string;
  message?: string;
}