# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  slug: string;
  name: string;
  priceCents: number;
  currency: string;
  paymentNetwork: string;
  tagline: string;
  description: string;
  features: string[];
}