# lib/api-zod/src/generated/types/reviewOrder.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Order } from './order';

export type ReviewOrder = Order & {
  userId: string;
  customerEmail: string;
  proofObjectPath?: string;
};